+++ AdminApi/Models/App/Bhaktas/Bhakta.cs
@@ -4,7 +4,7 @@ namespace AdminApi.Models.App.Bhaktas
 {
     public class Bhakta
     {
-        public Int64 BhaktaId { get; set; }
+        public int BhaktaId { get; set; }
         public string BhaktaName { get; set; }
         public string Address { get; set; }
         public string PrimaryPhone { get; set; }
